pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contentforge"
version = "0.1.0"
description = "CLI tool for generating content using LLMs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Silver" }]
keywords = ["cli", "content", "llm", "ai", "generation", "openai", "gemini", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "httpx>=0.25.0,<1.0.0",
    "openai>=1.0.0,<2.0.0",
    "google-generativeai>=0.4.0,<1.0.0",
    "tomli-w>=1.0.0,<2.0.0",
    "pyperclip>=1.8.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "pre-commit>=3.0",
    "build>=1.0",
]

[project.scripts]
contentforge = "contentforge.cli:app"

[project.urls]
Homepage = "https://github.com/brolyroly007/contentforge"
Repository = "https://github.com/brolyroly007/contentforge"
Issues = "https://github.com/brolyroly007/contentforge/issues"
Changelog = "https://github.com/brolyroly007/contentforge/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/contentforge"]

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
